+++ Sources/FrameUp/FULayout/AnyFULayout.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  AnyFULayout.swift
 //  
 //
 //  Created by Ryan Lintott on 2022-07-18.
@@ -10,8 +10,6 @@ import SwiftUI
 public struct AnyFULayout: FULayout {
     public let fuLayoutName: String
     public let id: UUID
-    public let alignment: Alignment
-    public let itemAlignment: Alignment
     public let fixedSize: Axis.Set
     public let maxItemWidth: CGFloat?
     public let maxItemHeight: CGFloat?
@@ -19,9 +17,7 @@ public struct AnyFULayout: FULayout {
     
     public init<L: FULayout>(_ layout: L) {
         fuLayoutName = layout.fuLayoutName
-        id = UUID() //layout.id
-        alignment = layout.alignment
-        itemAlignment = layout.itemAlignment
+        id = UUID()
         fixedSize = layout.fixedSize
         maxItemWidth = layout.maxItemWidth
         maxItemHeight = layout.maxItemHeight
@@ -35,8 +31,6 @@ public struct AnyFULayout: FULayout {
     public static func == (lhs: AnyFULayout, rhs: AnyFULayout) -> Bool {
         lhs.fuLayoutName == rhs.fuLayoutName
         && lhs.id == rhs.id
-        && lhs.alignment == rhs.alignment
-        && lhs.itemAlignment == rhs.itemAlignment
         && lhs.fixedSize == rhs.fixedSize
         && lhs.maxItemWidth == rhs.maxItemWidth
         && lhs.maxItemHeight == rhs.maxItemHeight
