+++ Sources/FrameUp/FULayout/FULayout+forEach.swift
@@ -1,5 +1,5 @@
 //
-//  FULayoutEach.swift
+//  FULayout+forEach.swift
 //  FrameUpExample
 //
 //  Created by Ryan Lintott on 2022-05-31.
